* sched_flags is also 8 bit now.

+++ modules/dhb/sched.h
@@ -4,7 +4,7 @@
 #  define _SCHED_H
 
 /* Scheduler state flags. */
-extern volatile unsigned int sched_flags; /**< Scheduler flags. */
+extern volatile uint8_t sched_flags; /**< Scheduler flags. */
 #define SCHED_HEARTBEAT             (1<<0) /**< HEARTBEAT Task. */
 #define SCHED_MOTOR                 (1<<1) /**< Motor control task. */
